+++ R/describe-feature.R
@@ -10,11 +10,18 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and limitations under the License.
 
-#' Describe a Web Service feature
+#' Describe the attributes of a Web Service feature
 #'
-#' Describe the columns from a Web Service feature. The column name, whether a column can be
-#' separated from the record in Web Service (nillable) and the type of column are returned.
-#' This can be a helpful tool to examine a layer before issuing a query with `bcdc_query_geodata`
+#' Describe the attributes of column of a record accessed through the Web Service.
+#' This can be a useful tool to examine a layer before issuing a query with `bcdc_query_geodata`.
+#'
+#' @return
+#' `bcdc_describe_features` returns a tibble describing the attributes of a B.C. Data Catalogue record.
+#' The tibble returns the following columns:
+#'   - col_name: attributes of the feature
+#'   - sticky: whether a column can be separated from the record in a Web Service call via the `dplyr::select` method
+#'   - remote_col_type: class of what is return by the web feature service
+#'   - local_col_type: the column class in R
 #'
 #' @inheritParams bcdc_query_geodata
 #' @export
@@ -52,7 +59,7 @@ bcdc_describe_feature.character <- function(record){
 bcdc_describe_feature.bcdc_record <- function(record){
 
   if (!any(wfs_available(record$resource_df))) {
-    stop("No WMS/WFS resource available for this dataset.",
+    stop("No WMS/WFS resource available for this data set.",
          call. = FALSE
     )
   }
@@ -63,12 +70,11 @@ bcdc_describe_feature.bcdc_record <- function(record){
 parse_raw_feature_tbl <- function(query_list){
 
   ## GET and parse data to sf object
-  cli <-
-    bcdc_http_client(url = "https://openmaps.gov.bc.ca/geo/pub/wfs")
+  cli <- bcdc_wfs_client()
 
   cc <- cli$post(body = query_list, encode = "form")
 
-  catch_catalogue_error(cc)
+  catch_wfs_error(cc)
 
   xml_res <- xml2::read_xml(cc$parse("UTF-8"))
   xml_res <- xml2::xml_find_all(xml_res, "//xsd:sequence")
@@ -110,7 +116,7 @@ feature_helper <- function(whse_name){
   xml_df <- xml_df[, c("name", "nillable", "type")]
   ## Add the id_row back into the front
   xml_df <- dplyr::bind_rows(id_row, xml_df)
-  colnames(xml_df) <- c("col_name", "selectable", "remote_col_type")
+  colnames(xml_df) <- c("col_name", "sticky", "remote_col_type")
   xml_df$local_col_type <- wfs_to_r_col_type(xml_df$remote_col_type)
 
   xml_df