+++ R/utils.R
@@ -10,7 +10,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and limitations under the License.
 
-base_url <- function() "https://catalogue.data.gov.bc.ca/api/3/"
+catalogue_base_url <- function() "https://catalogue.data.gov.bc.ca/api/3/"
+wfs_base_url <- function() "https://openmaps.gov.bc.ca/geo/pub/wfs/"
 
 bcdata_user_agent <- function(){
   "https://github.com/bcgov/bcdata"
@@ -21,13 +22,17 @@ compact <- function(l) Filter(Negate(is.null), l)
 
 bcdc_number_wfs_records <- function(query_list, client){
 
+  if (!is.null(query_list$count)) {
+    return(query_list$count)
+  }
+
   if(!is.null(query_list$propertyName)){
     query_list$propertyName <- NULL
   }
 
   query_list <- c(resultType = "hits", query_list)
   res_max <- client$post(body = query_list, encode = "form")
-  catch_catalogue_error(res_max)
+  catch_wfs_error(res_max)
   txt_max <- res_max$parse("UTF-8")
 
   ## resultType is only returned as XML.
@@ -74,11 +79,33 @@ formats_supported <- function(){
   c(bcdc_read_functions()[["format"]], "zip")
 }
 
-bcdc_http_client <- function(url = NULL) {
+bcdc_catalogue_client <- function(endpoint = NULL) {
+  url <- paste0(catalogue_base_url(), endpoint)
+  bcdc_http_client(url, auth = TRUE)
+}
 
-  crul::HttpClient$new(url = url,
-                       headers = list(`User-Agent` = bcdata_user_agent()))
+bcdc_wfs_client <- function(endpoint = NULL) {
+  url <- paste0(wfs_base_url(), endpoint)
+  bcdc_http_client(url, auth = FALSE)
+}
 
+bcdc_http_client <- function(url, auth = FALSE) {
+  headers <- list(
+    `User-Agent` = bcdata_user_agent(),
+    Authorization = if (auth) bcdc_auth() else NULL
+  )
+
+  crul::HttpClient$new(
+    url = url,
+    headers = compact(headers)
+  )
+}
+
+bcdc_auth <- function() {
+  key <- Sys.getenv("BCDC_KEY")
+  if (!nzchar(key)) return(NULL)
+  message("Authorizing with your stored API key")
+  key
 }
 
 ## Check if there is internet
@@ -124,7 +151,6 @@ wfs_to_r_col_type <- function(col){
     TRUE ~ as.character(col)
   )
 }
-
 
 ##from a record
 formats_from_record <- function(x, trim = TRUE){
@@ -197,8 +223,8 @@ read_from_url <- function(resource, ...){
   if (!reported_format %in% formats_supported()) {
     stop("Reading ", reported_format, " files is not currently supported in bcdata.")
   }
-
-  cli <- bcdc_http_client(file_url)
+  auth <- grepl("(catalogue|pub)\\.data\\.gov\\.bc\\.ca", file_url)
+  cli <- bcdc_http_client(file_url, auth = auth)
 
   ## Establish where to download file
   tmp <- tempfile(tmpdir = unique_temp_dir(),
@@ -220,10 +246,13 @@ read_from_url <- function(resource, ...){
   # where that's not the case
   message("Reading the data using the ", fun$fun, " function from the ",
           fun$package, " package.")
+  handle_excel(tmp, ...)
+
   tryCatch(do.call(fun$fun, list(tmp, ...)),
            error = function(e) {
-             stop("Could not read data set. The file can be found here:\n '",
-                  tmp, "'\n if you would like to try to read it manually.\n\n",
+             stop("Reading the data set failed with the following error message:\n\n  ", e,
+                  "\nThe file can be found here:\n  '",
+                  tmp, "'\nif you would like to try to read it manually.\n",
                   call. = FALSE)
            })
 }
@@ -281,6 +310,21 @@ handle_zip <- function(x) {
   files
 }
 
+handle_excel <- function(tmp, ...) {
+  if (!is_filetype(tmp, c("xls", "xlsx"))) {
+    return(invisible(NULL))
+  }
+
+  sheets <- readxl::excel_sheets(tmp)
+  if (length(sheets) > 1L) {
+    message(paste0("\nThis .", tools::file_ext(tmp), " resource contains the following sheets: \n",
+                   paste0(" '", sheets,"'", collapse = "\n")))
+    if (!methods::hasArg("sheet")) {
+      message("Defaulting to the '", sheets[1], "' sheet. See ?bcdc_get_data for examples on how to specify a sheet.\n")
+    }
+  }
+}
+
 
 unique_temp_dir <- function(pattern = "bcdata_") {
   dir <- tempfile(pattern = pattern)
@@ -304,27 +348,38 @@ list_supported_files <- function(dir) {
   files[supported]
 }
 
-catch_catalogue_error <- function(catalogue_response) {
-  msg <- "The BC data catalogue is currently unable to process this request\n"
+catch_wfs_error <- function(catalogue_response) {
+  msg <- "There was an issue sending this WFS request\n"
 
   if (inherits(catalogue_response, "Paginator")) {
     statuses <- catalogue_response$status_code()
     status_failed <- any(statuses >= 300)
-    msg <- paste0(msg, paste0(length(statuses), " paginated requests issued"))
+    if (!status_failed) return(invisible(NULL))
 
+    msg <- paste0(msg, paste0(length(statuses), " paginated requests issued"))
   } else {
     status_failed <- catalogue_response$status_code >= 300
+    if (!status_failed) return(invisible(NULL))
+
     request_res <- catalogue_response$request_headers
     response_res <- catalogue_response$response_headers
 
-    msg <- paste0(msg, "Catalogue request:\n")
+    msg <- paste0(
+      msg,
+      cli::rule(line = "bar4", line_col = 'red'),"\n",
+      "Request:",
+      "\n  URL: ", catalogue_response$request$url$url,
+      "\n  POST fields:\n    ", rawToChar(catalogue_response$request$options$postfields),
+      "\n"
+    )
+
     for (i in seq_along(request_res)) {
       msg <- paste0(
         msg, "  ", names(request_res)[i], ": ",
         request_res[i], "\n"
       )
     }
-    msg <- paste0(msg, "Catalogue response:\n")
+    msg <- paste0(msg, "Response:\n")
     for (i in seq_along(response_res)) {
       msg <- paste0(
         msg, "  ", names(response_res)[i], ": ",
@@ -333,7 +388,5 @@ catch_catalogue_error <- function(catalogue_response) {
     }
   }
 
-  if (status_failed) {
-    stop(msg, call. = FALSE)
-  }
+  stop(msg, call. = FALSE)
 }