+++ tests/testthat/test-query-geodata-collect.R
@@ -20,6 +20,14 @@ test_that("bcdc_query_geodata collects an sf object for a valid id", {
   expect_equal(attr(bc_airports, "sf_column"), "geometry")
 })
 
+test_that("bcdc_query_geodata collects using as_tibble", {
+  skip_on_cran()
+  skip_if_net_down()
+  bc_airports <- bcdc_query_geodata("bc-airports") %>% as_tibble()
+  expect_is(bc_airports, "sf")
+  expect_equal(attr(bc_airports, "sf_column"), "geometry")
+})
+
 test_that("bcdc_query_geodata succeeds with a records over 10000 rows",{
   skip_on_cran()
   skip("Skipping the BEC test, though available for testing")
@@ -40,9 +48,14 @@ test_that("bcdc_query_geodata works with slug and full url using collect", {
             "sf")
   expect_is(ret4 <- bcdc_query_geodata("76b1b7a3-2112-4444-857a-afccf7b20da8") %>% collect(),
             "sf")
-  ## Must be a better way to test if these objects are equal
-  expect_true(all(unlist(lapply(list(ret1, ret2, ret3, ret4), nrow))))
-  expect_true(all(unlist(lapply(list(ret1, ret2, ret3, ret4), ncol))))
+  expect_is(ret5 <- bcdc_query_geodata("https://catalogue.data.gov.bc.ca/dataset/76b1b7a3-2112-4444-857a-afccf7b20da8/resource/4d0377d9-e8a1-429b-824f-0ce8f363512c")
+            %>% collect(),
+            "sf")
+
+  for (x in list(ret2, ret3, ret4, ret5)) {
+    expect_equal(dim(x), dim(ret1))
+    expect_equal(names(x), names(ret1))
+  }
 })
 
 
@@ -73,7 +86,7 @@ test_that("collect() returns a bcdc_sf object",{
   expect_s3_class(sf_obj, "bcdc_sf")
 })
 
-test_that("bcdc_sf objects have a url as an attributes",{
+test_that("bcdc_sf objects has attributes",{
   skip_on_cran()
   skip_if_net_down()
   sf_obj <- bcdc_query_geodata("76b1b7a3-2112-4444-857a-afccf7b20da8") %>%
@@ -81,5 +94,10 @@ test_that("bcdc_sf objects have a url as an attributes",{
     select(LATITUDE) %>%
     collect()
 
+  expect_identical(names(attributes(sf_obj)),
+                   c("names", "row.names", "class", "sf_column", "agr", "query_list",
+                     "url", "full_url", "time_downloaded"))
   expect_true(nzchar(attributes(sf_obj)$url))
+  expect_true(nzchar(attributes(sf_obj)$full_url))
+  expect_is(attributes(sf_obj)$time_downloaded, "POSIXt")
 })